requirements.txt
discord.py
python-dotenv
aiohttp
python-telegram-bot
aiofiles
